+++ src/main/scala/io/fintrospect/parameters/HeaderParameter.scala
@@ -1,30 +1,46 @@
 package io.fintrospect.parameters
 
-import com.twitter.finagle.http.Message
+import com.twitter.finagle.http.{Message, Request}
 
 import scala.util.{Failure, Success, Try}
 
-abstract class HeaderParameter[T](spec: ParameterSpec[T])
+abstract class HeaderParameter[T](spec: ParameterSpec[_], val deserialize: Seq[String] => T)
   extends Parameter
   with Validatable[T, Message]
   with Bindable[T, RequestBinding] {
   override val name = spec.name
   override val description = spec.description
   override val paramType = spec.paramType
-
-  override def -->(value: T) = Seq(new RequestBinding(this, in => {
-    in.headerMap.add(name, spec.serialize(value))
-    in
-  }))
-
   val where = "header"
 
   def validate(message: Message) = {
-    message.headerMap.get(name).map {
-      v => Try(spec.deserialize(v)) match {
-        case Success(d) => Right(Option(d))
-        case Failure(_) => Left(this)
-      }
-    }.getOrElse(if (required) Left(this) else Right(None))
+    val headers = message.headerMap.getAll(name)
+    val opt = if(headers.isEmpty) None else Some(headers.toSeq)
+    opt
+      .map(v => Try(deserialize(v)) match {
+          case Success(d) => Right(Option(d))
+          case Failure(_) => Left(this)
+        }).getOrElse(if (required) Left(this) else Right(None))
   }
 }
+
+abstract class SingleHeaderParameter[T](spec: ParameterSpec[T])
+  extends HeaderParameter[T](spec, xs => spec.deserialize(xs.head)) {
+  override def -->(value: T) = Seq(new RequestBinding(this, {
+    req: Request => {
+      req.headerMap.add(spec.name, spec.serialize(value))
+      req
+    }
+  }))
+}
+
+abstract class MultiHeaderParameter[T](spec: ParameterSpec[T])
+  extends HeaderParameter[Seq[T]](spec, xs => xs.map(spec.deserialize)) {
+  override def -->(value: Seq[T]) = value.map(v => new RequestBinding(this, {
+    req: Request => {
+      req.headerMap.add(spec.name, spec.serialize(v))
+      req
+    }
+  }))
+}
+