+++ src/test/scala/examples/customformats/HipsterXmlApp.scala
@@ -10,6 +10,7 @@ import com.twitter.finagle.{Http, Service}
 import examples.customformats.HipsterXml.ResponseBuilder.implicits.{responseBuilderToFuture, statusToResponseBuilderConfig}
 import io.fintrospect.parameters.{Body, BodySpec, Header, ParameterSpec, Path, StringParamType}
 import io.fintrospect.{ContentTypes, ModuleSpec, RouteSpec}
+import scala.language.reflectiveCalls
 
 /**
  * This application shows how to define a custom rendering, body and parameter formats.